Add WIP git message template
Why:

Sometimes the standard message template isn't the right thing, e.g.
when many changes are split across several templates. In that case it's
more useful to use a WIP template with information about what's happened
so far and what's next, rather than the why/how sections in the default
template. The latter are more suited to smaller, self-contained commits.

This change addresses the need by:

* Creating a new WIP git message template
* Adding an alias to use it (`gciwip`)
* Also improve comments in functions file.