Don't use ruby prompt in non-interactive shells
Why this change is necessary:

Some installers for applications run shell scripts as part of the installation. For a while, I had been noticing that certain applications' (usually those from Apple) installation would stall. Checking running processes, this was usually due to an invocation of `bash rbenv-gemset-file`. After some investigation, I figured out that the above command was being invoked from `rbenv gemset active`, which in turn was invoked from the parts of my prompt that check for the active gemset.

This change addresses the need by:

Invoking `rbenv gemset active`, and actually setting the ruby prompt, only if the shell is interactive.