Get vi mode indicator working again.
This change is necessary because:

At some point the `VI_MODE_RESET_PROMPT_ON_MODE_CHANGE` env var was introduced, which if unset causes the vi-mode zsh plugin not to redraw the prompt on input mode change. As a result, the prompt wasn't updating to show when the prompt was in normal mode.

This change addresses the need by:

* Setting `VI_MODE_RESET_PROMPT_ON_MODE_CHANGE` to true in the zsh theme to ensure that the prompt updates on mode change.
* Also includes a vim config to properly highlight `*.zsh-theme` files